refactor: make getHibernatedPVCGroup reusable
